Link AppUser to Applications

+++ api/Controllers/ApplicationController.cs
@@ -4,10 +4,13 @@ using System.Linq;
 using System.Threading.Tasks;
 using api.Data;
 using api.Dtos.Application;
+using api.Extensions;
 using api.Helpers;
 using api.Interfaces;
 using api.Mappers;
 using api.Models;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 
 
@@ -19,10 +22,12 @@ namespace api.Controllers
     {
         public readonly ApplicationDBContext _context;
         public readonly IApplicationRepository _applicationRepo;
-        public ApplicationController(ApplicationDBContext context, IApplicationRepository applicationRepository)
+        private readonly UserManager<AppUser> _userManager;
+        public ApplicationController(ApplicationDBContext context, IApplicationRepository applicationRepository, UserManager<AppUser> userManager)
         {
             _context = context;
             _applicationRepo = applicationRepository;
+            _userManager = userManager;
         }
 
         [HttpGet]
@@ -33,9 +38,17 @@ namespace api.Controllers
                 return BadRequest(ModelState);
             }
 
-            var applications = await _applicationRepo.GetAllAsync(query);
+            var username = User.GetUserName();
+            var appUser = await _userManager.FindByNameAsync(username);
+
+            if (appUser == null)
+            {
+                return Unauthorized(); // Ensure the user is logged in
+            }
+
+            var applications = await _applicationRepo.GetAllAsync(query, appUser.Id);
             var applicationDto = applications.Select(s => s.ToApplicationDto());
-            return Ok(applications);
+            return Ok(applicationDto);
         }
 
         [HttpGet("{id:int}")]
@@ -46,7 +59,15 @@ namespace api.Controllers
                 return BadRequest(ModelState);
             }
 
-            var application = await _applicationRepo.GetByIdAsync(id);
+            var username = User.GetUserName();
+            var appUser = await _userManager.FindByNameAsync(username);
+
+            if (appUser == null)
+            {
+                return Unauthorized(); // Ensure the user is logged in
+            }
+
+            var application = await _applicationRepo.GetByIdAsync(id, appUser.Id);
             if (application == null)
             {
                 return NotFound();
@@ -62,7 +83,16 @@ namespace api.Controllers
                 return BadRequest(ModelState);
             }
 
+            var username = User.GetUserName();
+            var appUser = await _userManager.FindByNameAsync(username);
+
+            if (appUser == null)
+            {
+                return Unauthorized(); // Ensure the user is logged in
+            }
+
             var application = applicationDto.ToApplicationFromCreateDto();
+            application.UserId = appUser.Id;
             await _applicationRepo.CreateAsync(application);
             return CreatedAtAction(nameof(GetById), new { id = application.Id }, application.ToApplicationDto());
         }
@@ -75,8 +105,16 @@ namespace api.Controllers
                 return BadRequest(ModelState);
             }
 
+            var username = User.GetUserName();
+            var appUser = await _userManager.FindByNameAsync(username);
+
+            if (appUser == null)
+            {
+                return Unauthorized(); // Ensure the user is logged in
+            }
+
             var application = await _applicationRepo.UpdateAsync(id, applicationDto);
-            if (application == null)
+            if (application == null || application.UserId != appUser.Id)
             {
                 return NotFound();
             }
@@ -91,8 +129,17 @@ namespace api.Controllers
             {
                 return BadRequest(ModelState);
             }
+
+            var username = User.GetUserName();
+            var appUser = await _userManager.FindByNameAsync(username);
+
+            if (appUser == null)
+            {
+                return Unauthorized(); // Ensure the user is logged in
+            }
+
             var application = await _applicationRepo.DeleteAsync(id);
-            if (application == null)
+            if (application == null || application.UserId != appUser.Id)
             {
                 return NotFound();
             }
